% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_owl.R
\name{read_owl}
\alias{read_owl}
\title{Read an OWL File}
\usage{
read_owl(owl_path)
}
\arguments{
\item{owl_path}{Path to OWL file.}
}
\value{
List Object
}
\description{
Runs the \code{xmltodict.parse} in python on an OWL file.
}
\examples{
library(owlr)
test_owl <-
  read_owl(
    file.path(
      system.file(
        package = "owlr", "owl"),
      "test.owl")
  )
test_owl
}
